# Array Encoding/Decoding performance improvements (continued)