--- a/src/Encoding/BinaryEncoding.cs
+++ b/src/Encoding/BinaryEncoding.cs
@@ -4,8 +4,9 @@
 namespace Microsoft.Azure.Amqp.Encoding
 {
     using System;
+    using System.Collections.Generic;
 
-    sealed class BinaryEncoding : EncodingBase
+    sealed class BinaryEncoding : PrimitiveEncoding<ArraySegment<byte>>
     {
         public BinaryEncoding()
             : base(FormatCode.Binary32)
@@ -39,7 +40,7 @@ namespace Microsoft.Azure.Amqp.Encoding
                     AmqpBitConverter.WriteUInt(buffer, (uint)value.Count);
                 }
 
-                AmqpBitConverter.WriteBytes(buffer, value.Array, value.Offset, value.Count);
+                AmqpBitConverter.WriteBytes(buffer, value, value.Offset, value.Count);
             }
         }
 
@@ -61,23 +62,21 @@ namespace Microsoft.Azure.Amqp.Encoding
             {
                 return AmqpConstants.EmptyBinary;
             }
+
+            ArraySegment<byte> value;
+            if (copy)
+            {
+                byte[] valueBuffer = new byte[count];
+                Buffer.BlockCopy(buffer.Buffer, buffer.Offset, valueBuffer, 0, count);
+                value = new ArraySegment<byte>(valueBuffer, 0, count);
+            }
             else
             {
-                ArraySegment<byte> value;
-                if (copy)
-                {
-                    byte[] valueBuffer = new byte[count];
-                    Buffer.BlockCopy(buffer.Buffer, buffer.Offset, valueBuffer, 0, count);
-                    value = new ArraySegment<byte>(valueBuffer, 0, count);
-                }
-                else
-                {
-                    value = new ArraySegment<byte>(buffer.Buffer, buffer.Offset, count);
-                }
-
-                buffer.Complete(count);
-                return value;
+                value = new ArraySegment<byte>(buffer.Buffer, buffer.Offset, count);
             }
+
+            buffer.Complete(count);
+            return value;
         }
 
         public override int GetObjectEncodeSize(object value, bool arrayEncoding)
@@ -86,10 +85,8 @@ namespace Microsoft.Azure.Amqp.Encoding
             {
                 return FixedWidth.UInt + ((ArraySegment<byte>)value).Count;
             }
-            else
-            {
-                return BinaryEncoding.GetEncodeSize((ArraySegment<byte>)value);
-            }
+
+            return BinaryEncoding.GetEncodeSize((ArraySegment<byte>)value);
         }
 
         public override void EncodeObject(object value, bool arrayEncoding, ByteBuffer buffer)
@@ -98,7 +95,7 @@ namespace Microsoft.Azure.Amqp.Encoding
             {
                 ArraySegment<byte> binaryValue = (ArraySegment<byte>)value;
                 AmqpBitConverter.WriteUInt(buffer, (uint)binaryValue.Count);
-                AmqpBitConverter.WriteBytes(buffer, binaryValue.Array, binaryValue.Offset, binaryValue.Count);
+                AmqpBitConverter.WriteBytes(buffer, binaryValue, binaryValue.Offset, binaryValue.Count);
             }
             else
             {
@@ -109,6 +106,55 @@ namespace Microsoft.Azure.Amqp.Encoding
         public override object DecodeObject(ByteBuffer buffer, FormatCode formatCode)
         {
             return BinaryEncoding.Decode(buffer, formatCode, false);
+        }
+
+        public override int GetArrayEncodeSize(IList<ArraySegment<byte>> value)
+        {
+            int size = 0;
+            for (int i = 0; i < value.Count; i++)
+            {
+                size += AmqpEncoding.GetEncodeWidthBySize(value[i].Count);
+            }
+            return size;
+        }
+
+        public override void EncodeArray(IList<ArraySegment<byte>> value, ByteBuffer buffer)
+        {
+            if (value is ArraySegment<byte>[] arraySegments)
+            {
+                // fast-path for ArraySegment<byte>[] so the bounds checks can be elided
+                for (int i = 0; i < arraySegments.Length; i++)
+                {
+                    var segment = arraySegments[i];
+                    AmqpBitConverter.WriteUInt(buffer, (uint)segment.Count);
+                    AmqpBitConverter.WriteBytes(buffer, segment, segment.Offset, segment.Count);
+                }
+            }
+            else
+            {
+                IReadOnlyList<ArraySegment<byte>> listValue = (IReadOnlyList<ArraySegment<byte>>)value;
+                for (int i = 0; i < listValue.Count; i++)
+                {
+                    var segment = listValue[i];
+                    AmqpBitConverter.WriteUInt(buffer, (uint)segment.Count);
+                    AmqpBitConverter.WriteBytes(buffer, segment, segment.Offset, segment.Count);
+                }
+            }
+        }
+
+        public override ArraySegment<byte>[] DecodeArray(ByteBuffer buffer, int count, FormatCode formatCode)
+        {
+            ArraySegment<byte>[] array = new ArraySegment<byte>[count];
+            for (int i = 0; i < count; ++i)
+            {
+                AmqpEncoding.ReadCount(buffer, formatCode, FormatCode.Binary8, FormatCode.Binary32, out count);
+                // always copy?
+                byte[] valueBuffer = new byte[count];
+                var arraySegment = new ArraySegment<byte>(valueBuffer, 0, count);
+                AmqpBitConverter.ReadBytes(buffer, arraySegment, 0, count);
+                array[i] = arraySegment;
+            }
+            return array;
         }
     }
 }