--- a/src/Encoding/DecimalEncoding.cs
+++ b/src/Encoding/DecimalEncoding.cs
@@ -4,7 +4,6 @@
 namespace Microsoft.Azure.Amqp.Encoding
 {
     using System;
-    using System.Buffers.Binary;
     using System.Collections.Generic;
 
     /// <summary>