… src/Encoding/CharEncoding.cs
… src/Encoding/DoubleEncoding.cs
… src/Encoding/FloatEncoding.cs
… src/Encoding/IntEncoding.cs
… src/Encoding/LongEncoding.cs
… src/Encoding/ShortEncoding.cs
… src/Encoding/StringEncoding.cs
… src/Encoding/SymbolEncoding.cs
… src/Encoding/TimeStampEncoding.cs
… src/Encoding/UByteEncoding.cs
… src/Encoding/UIntEncoding.cs
… src/Encoding/ULongEncoding.cs
… src/Encoding/UShortEncoding.cs
… src/Encoding/UuidEncoding.cs
… src/Framing/Multiple.cs
… src/Microsoft.Azure.Amqp.csproj
… test/TestCases/AmqpCodecTests.cs