@danielmarbach 2021-06-08
@xinchen10 I think I'm missing one primitive here and running the latest results in the benchmark (which is still quite a bit of effort) 

Can you have a look and tell me what you think or if there is anything missing from your side?